# sacdia-api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/clubs": {
      "get": {
        "description": "Obtiene la lista de clubs con filtros opcionales y paginación",
        "operationId": "ClubsController_findAll_v1",
        "parameters": [
          {
            "name": "localFieldId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "districtId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "churchId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de página (1-indexed)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Elementos por página (max 100)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de clubs"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar clubs",
        "tags": [
          "clubs"
        ]
      },
      "post": {
        "operationId": "ClubsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClubDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Club creado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear nuevo club",
        "tags": [
          "clubs"
        ]
      }
    },
    "/v1/clubs/{clubId}": {
      "get": {
        "operationId": "ClubsController_findOne_v1",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Club encontrado"
          },
          "404": {
            "description": "Club no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener club por ID",
        "tags": [
          "clubs"
        ]
      },
      "patch": {
        "operationId": "ClubsController_update_v1",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClubDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Club actualizado"
          },
          "403": {
            "description": "Permisos insuficientes"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar club (requiere rol director o subdirector)",
        "tags": [
          "clubs"
        ]
      },
      "delete": {
        "operationId": "ClubsController_remove_v1",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Club desactivado"
          },
          "403": {
            "description": "Permisos insuficientes"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desactivar club (requiere rol director)",
        "tags": [
          "clubs"
        ]
      }
    },
    "/v1/clubs/{clubId}/instances": {
      "get": {
        "description": "Lista todas las instancias (Aventureros, Conquistadores, GM)",
        "operationId": "ClubsController_getInstances_v1",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Instancias del club"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener instancias del club",
        "tags": [
          "clubs"
        ]
      },
      "post": {
        "description": "Crea una nueva instancia (Aventureros, Conquistadores, Guías Mayores)",
        "operationId": "ClubsController_createInstance_v1",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Instancia creada"
          },
          "403": {
            "description": "Permisos insuficientes"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear instancia de club (requiere director o subdirector)",
        "tags": [
          "clubs"
        ]
      }
    },
    "/v1/clubs/{clubId}/instances/{type}": {
      "get": {
        "operationId": "ClubsController_getInstance_v1",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "path",
            "description": "Tipo de instancia",
            "schema": {
              "enum": [
                "adventurers",
                "pathfinders",
                "master_guilds"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Instancia encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener instancia por tipo",
        "tags": [
          "clubs"
        ]
      }
    },
    "/v1/clubs/{clubId}/instances/{type}/{instanceId}": {
      "patch": {
        "operationId": "ClubsController_updateInstance_v1",
        "parameters": [
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "adventurers",
                "pathfinders",
                "master_guilds"
              ],
              "type": "string"
            }
          },
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInstanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Instancia actualizada"
          },
          "403": {
            "description": "Permisos insuficientes"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar instancia (requiere director, subdirector o secretario)",
        "tags": [
          "clubs"
        ]
      }
    },
    "/v1/clubs/{clubId}/instances/{type}/{instanceId}/members": {
      "get": {
        "description": "Retorna todos los miembros asignados a la instancia con sus roles",
        "operationId": "ClubsController_getMembers_v1",
        "parameters": [
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "adventurers",
                "pathfinders",
                "master_guilds"
              ],
              "type": "string"
            }
          },
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de miembros"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar miembros de la instancia",
        "tags": [
          "clubs"
        ]
      }
    },
    "/v1/clubs/{clubId}/instances/{type}/{instanceId}/roles": {
      "post": {
        "description": "Asigna un rol específico a un usuario en la instancia",
        "operationId": "ClubsController_assignRole_v1",
        "parameters": [
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "adventurers",
                "pathfinders",
                "master_guilds"
              ],
              "type": "string"
            }
          },
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rol asignado"
          },
          "403": {
            "description": "Permisos insuficientes"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Asignar rol a un miembro (requiere director, subdirector o secretario)",
        "tags": [
          "clubs"
        ]
      }
    },
    "/v1/club-roles/{assignmentId}": {
      "patch": {
        "operationId": "ClubRolesController_updateAssignment_v1",
        "parameters": [
          {
            "name": "assignmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Asignación actualizada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar asignación de rol",
        "tags": [
          "club-roles"
        ]
      },
      "delete": {
        "operationId": "ClubRolesController_removeAssignment_v1",
        "parameters": [
          {
            "name": "assignmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rol removido"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover rol de miembro",
        "tags": [
          "club-roles"
        ]
      }
    },
...
  },
  "info": {
    "title": "SACDIA API",
    "description": "## Sistema de Administración de Clubes de Conquistadores y Aventureros\n\n### Módulos Disponibles\n- **Auth**: Autenticación con Supabase + JWT\n- **Users**: Gestión de perfiles de usuario\n- **Catalogs**: Catálogos de referencia\n- **Clubs**: Gestión de clubes e instancias\n- **Classes**: Clases progresivas\n- **Honors**: Especialidades\n- **Activities**: Actividades de club\n- **Finances**: Control financiero",
    "version": "2.2.0",
    "contact": {
      "name": "SACDIA Team",
      "url": "https://sacdia.app",
      "email": "dev@sacdia.app"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateClubDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "example": "Club Central",
            "description": "Nombre del club"
          },
          "description": {
            "type": "string",
            "description": "Descripción del club"
          },
          "local_field_id": {
            "type": "number",
            "example": 1,
            "description": "ID del campo local"
          },
          "districlub_type_id": {
            "type": "number",
            "example": 1,
            "description": "ID del distrito"
          },
          "church_id": {
            "type": "number",
            "example": 1,
            "description": "ID de la iglesia"
          },
          "address": {
            "type": "string",
            "description": "Dirección del club"
          },
          "coordinates": {
            "type": "object",
            "properties": {
              "lat": {
                "type": "number"
              },
              "lng": {
                "type": "number"
              }
            },
            "required": [
              "lat",
              "lng"
            ]
          }
        },
        "required": [
          "name",
          "local_field_id",
          "districlub_type_id",
          "church_id"
        ]
      },
      "UpdateClubDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "example": "Club Actualizado"
          },
          "description": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "coordinates": {
            "type": "object",
            "properties": {
              "lat": {
                "type": "number"
              },
              "lng": {
                "type": "number"
              }
            },
            "required": [
              "lat",
              "lng"
            ]
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "CreateInstanceDto": {
        "type": "object",
        "properties": {
          "type": {
            "enum": [
              "adventurers",
              "pathfinders",
              "master_guilds"
            ],
            "type": "string",
            "example": "pathfinders",
            "description": "Tipo de instancia a crear"
          },
          "souls_target": {
            "type": "number",
            "example": 1,
            "description": "Meta de almas"
          },
          "fee": {
            "type": "number",
            "example": 100,
            "description": "Cuota mensual"
          },
          "meeting_day": {
            "example": [
              {
                "day": "Saturday"
              }
            ],
            "description": "Días de reunión",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meeting_time": {
            "example": [
              {
                "time": "09:00"
              }
            ],
            "description": "Horarios de reunión",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "type"
        ]
      },
      "UpdateInstanceDto": {
        "type": "object",
        "properties": {
          "souls_target": {
            "type": "number"
          },
          "fee": {
            "type": "number"
          },
          "meeting_day": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meeting_time": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "AssignRoleDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID del usuario"
          },
          "role_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID del rol a asignar"
          },
          "instance_type": {
            "enum": [
              "adventurers",
              "pathfinders",
              "master_guilds"
            ],
            "type": "string",
            "description": "Tipo de instancia de club"
          },
          "instance_id": {
            "type": "number",
            "description": "ID de la instancia de club"
          },
          "ecclesiastical_year_id": {
            "type": "number",
            "description": "ID del año eclesiástico"
          },
          "start_date": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de inicio del rol"
          },
          "end_date": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de fin del rol"
          }
        },
        "required": [
          "user_id",
          "role_id",
          "instance_type",
          "instance_id",
          "ecclesiastical_year_id",
          "start_date"
        ]
      },
      "UpdateRoleAssignmentDto": {
        "type": "object",
        "properties": {
          "end_date": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de fin del rol"
          },
          "status": {
            "type": "string",
            "description": "Estado del rol"
          }
        }
      },
...
    }
  }
}